Decouple onDragEnd handler by moving to each individual draggable component
